Move arapp.ts to main utils folder
