+++ src/tasks/start/backend/backend.ts
@@ -7,7 +7,7 @@ import { resolveRepo, majorBumpRepo } from './repo'
 import { setAllPermissionsOpenly } from './permissions'
 import { KernelInstance, RepoInstance } from '~/typechain'
 import { logBack } from '../utils/logger'
-import { readArapp } from '../utils/arapp'
+import { readArapp } from '../../../utils/arappUtils'
 import { AragonConfig, AragonConfigHooks } from '~/src/types'
 import { TASK_COMPILE } from '../../task-names'
 import deployAragonBases from './bases'
