+++ src/tasks/start/backend/repo.ts
@@ -1,3 +1,6 @@
+import Web3 from 'web3'
+import { TruffleEnvironmentArtifacts } from '@nomiclabs/buidler-truffle5/src/artifacts'
+import { getLog } from '../../../utils/getLog'
 import {
   RepoContract,
   RepoInstance,
@@ -6,9 +9,6 @@ import {
   /* PublicResolverContract, */
   /* PublicResolverInstance */
 } from '~/typechain'
-import Web3 from 'web3'
-import { TruffleEnvironmentArtifacts } from '@nomiclabs/buidler-truffle5/src/artifacts'
-import { getLog } from '../../../utils/getLog'
 
 /* const ZERO_ADDR = '0x0000000000000000000000000000000000000000' */
