+++ test/tasks/start/backend/permissions.test.ts
@@ -1,9 +1,11 @@
 import { assert } from 'chai'
 import { createDao, createDaoFactory } from '~/src/tasks/start/backend/dao'
-import { readArapp } from '~/src/tasks/start/utils/arapp'
+import { readArapp } from '~/src/utils/arappUtils'
 import { deployImplementation } from '~/src/tasks/start/backend/app'
 import { createProxy } from '~/src/tasks/start/backend/proxy'
 import { AragonAppJson } from '~/src/types'
+import { getAppId } from '~/src/tasks/start/utils/id'
+import { useDefaultEnvironment } from '~/test/test-helpers/useEnvironment'
 import {
   setAllPermissionsOpenly,
   ANY_ADDRESS
@@ -14,8 +16,6 @@ import {
   ACLInstance,
   DAOFactoryInstance
 } from '~/typechain'
-import { getAppId } from '~/src/tasks/start/utils/id'
-import { useDefaultEnvironment } from '~/test/test-helpers/useEnvironment'
 
 describe('permissions.ts', function() {
   useDefaultEnvironment()
