+++ test/tasks/start/backend/proxy.test.ts
@@ -2,10 +2,9 @@ import { assert } from 'chai'
 import { useDefaultEnvironment } from '~/test/test-helpers/useEnvironment'
 import { deployImplementation } from '~/src/tasks/start/backend/app'
 import { createProxy, updateProxy } from '~/src/tasks/start/backend/proxy'
-import { getAppEnsName } from '~/src/tasks/start/utils/arapp'
+import { readArapp, getAppEnsName } from '~/src/utils/arappUtils'
 import { getAppId } from '~/src/tasks/start/utils/id'
 import { setAllPermissionsOpenly } from '~/src/tasks/start/backend/permissions'
-import { readArapp } from '~/src/tasks/start/utils/arapp'
 import { createDao, createDaoFactory } from '~/src/tasks/start/backend/dao'
 
 describe('proxy.ts', function() {
